"ads2_m6_a1_1: 4/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS - 2 Assignments/m6/ADS-2Week-1/Solution.java
@@ -20,7 +20,7 @@ class PageRank {
 		}
 		graph2 = graph.reverse();
 		for (int i = 0; i < pRank.length; i++) {
-			pRank[i] = 1 / graph.V();
+			pRank[i] = 1 / (double)graph.V();
 		}
 		double[] temparr = new double[graph.V()];
 		for (int i = 0; i < 1000; i++) {
